# Conflicts:
#	src/main/java/com/tenji/adminapi2/dto/GrantedHolidayForm.java
#	src/main/java/com/tenji/adminapi2/mapper/GrantedHolidayMapper.java
#	src/main/java/com/tenji/adminapi2/model/GrantedHoliday.java
#	src/main/java/com/tenji/adminapi2/service/Impl/GrantedHolidayServiceImpl.java
#	src/main/resources/mapper/GrantedHolidayMapper.xml

+++ src/main/java/com/tenji/adminapi2/dto/GrantedHolidayForm.java
@@ -11,7 +11,7 @@ public class GrantedHolidayForm {
     //管理者ID
   //  private long userId;
     //社員ID
-    private Integer employeeId;
+    private long employeeId;
     //付与時社歴
  //   private float grantedServiceYears;
     //付与年
